+++ src/egeo-demo/st-tag-input-demo/st-tag-input-demo.component.ts
@@ -10,9 +10,8 @@
  */
 import { Component, ViewChild, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, Validators, NgForm } from '@angular/forms';
-
 import { cloneDeep as _cloneDeep } from 'lodash';
-import { StDropDownMenuItem } from '@stratio/egeo';
+import { StDropDownMenuItem, StDropDownMenuGroup } from '@stratio/egeo';
 
 @Component({
    selector: 'st-tag-input-demo',
@@ -30,7 +29,8 @@ export class StTagInputDemoComponent implements OnInit {
       templateDriven: ['Tag Template Driven1', 'Tag Template Driven2'],
       templateDrivenRequired: [],
       templateDrivenDisabled: [],
-      templateDrivenAutocomplete: []
+      templateDrivenAutocomplete: [],
+      autoCompletedGroupList: []
    };
 
    public list: StDropDownMenuItem[] = [
@@ -48,7 +48,49 @@ export class StTagInputDemoComponent implements OnInit {
       { label: 'Vanuatu', value: 'VU' },
       { label: 'Venezuela', value: 'VE' }
    ];
-   public filteredlist: StDropDownMenuItem[] = [];
+   public filteredList: StDropDownMenuItem[] = [];
+   public groupList: StDropDownMenuGroup[] = [
+      <StDropDownMenuGroup> {
+         title: 'Group1', items: [
+            { label: 'Option 1.1', value: 'option11' },
+            { label: 'Option 1.2', value: 'option12' }
+         ]
+      },
+      <StDropDownMenuGroup> {
+         title: 'Group2', items: [
+            { label: 'Option 2.1', value: 'option21' },
+            { label: 'Option 2.2', value: 'option22' },
+            { label: 'Option 2.3', value: 'option23' },
+            { label: 'Option 2.4', value: 'option24' }
+         ]
+      },
+      <StDropDownMenuGroup> {
+         title: 'Group3', items: [
+            { label: 'Option 3.1', value: 'option31' },
+            { label: 'Option 3.2', value: 'option32' }
+         ]
+      },
+      <StDropDownMenuGroup> {
+         title: 'Group4', items: [
+            { label: 'Option 4.1', value: 'option41' },
+            { label: 'Option 4.2', value: 'option42' },
+            { label: 'Option 4.3', value: 'option43' }
+         ]
+      },
+      <StDropDownMenuGroup> {
+         title: 'Group5', items: [
+            { label: 'Option 5.1', value: 'option51' },
+            { label: 'Option 5.2', value: 'option52' }
+         ]
+      },
+      <StDropDownMenuGroup> {
+         title: 'Group6', items: [
+            { label: 'Option 6.1', value: 'option61' },
+            { label: 'Option 6.2', value: 'option62' }
+         ]
+      }
+   ];
+   public filteredGroupList: StDropDownMenuGroup[] = [];
 
    public reactiveForm: FormGroup; // our model driven form
    public forceReactiveValidations: boolean = false;
@@ -57,7 +99,7 @@ export class StTagInputDemoComponent implements OnInit {
    public errorTemplateDriveMessage: string | null = null;
    public disabledReactive: boolean = true;
    public disabledTemplateDrive: boolean = true;
-   public pattern: any =  /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/;
+   public pattern: any = /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/;
 
    constructor(private _fb: FormBuilder) {
       this.reactiveForm = _fb.group({
@@ -66,13 +108,15 @@ export class StTagInputDemoComponent implements OnInit {
          'tag-input-integer': [],
          'tag-input-reactive-required': [this.tags.reactiveRequired, Validators.required],
          'tag-input-reactive-disabled': [this.tags.reactiveDisabled],
-         'tag-input-reactive-autocomplete': [this.tags.reactiveDisabled]
+         'tag-input-reactive-autocomplete': [this.tags.reactiveDisabled],
+         'tag-input-auto-completed-group-list': [this.tags.autoCompletedGroupList, Validators.compose([Validators.minLength(2), Validators.maxLength(5)])],
       });
    }
 
    ngOnInit(): void {
       this.reactiveForm.valueChanges.subscribe(res => console.log('Reactive Form', res));
       this.templateDrivenForm.valueChanges.subscribe(res => console.log('Template Driven Form', res));
+      this.filteredGroupList = _cloneDeep(this.groupList);
    }
 
    onSubmitReactiveForm(): void {
@@ -105,6 +149,31 @@ export class StTagInputDemoComponent implements OnInit {
 
    onFilterList(event: any): void {
       let text: string = event.target.innerText;
-      this.filteredlist = text ? _cloneDeep(this.list.filter(country => country.label.toLowerCase().search(text.toLowerCase()) > -1)) : [];
+      this.filteredList = text ? _cloneDeep(this.list.filter(country => country.label.toLowerCase().indexOf(text.toLowerCase()) !== -1)) : [];
+   }
+
+   onFilterGroupList(event: any): void {
+      let text: string = event.target.innerText;
+      console.log('voy a buscar con ' + text);
+      this.filteredGroupList = _cloneDeep(this.groupList);
+      this.filteredGroupList = <StDropDownMenuGroup[]> this.filteredGroupList.filter(item => {
+         const result = item.items.filter(_item => {
+            return _item.label.toLowerCase().indexOf(text.toLowerCase()) !== -1;
+         });
+         item.items = result;
+         return result && result.length ? item : undefined;
+      });
+      console.log(this.filteredGroupList);
+   }
+
+   getGroupAutoCompletedErrorMessage(): string {
+      const currentValue = this.reactiveForm.controls['tag-input-auto-completed-group-list'].value;
+      if (currentValue.length < 2) {
+         return 'You have to enter two items at leas';
+      }
+      else if (currentValue.length > 5) {
+         return 'Only 5 items are allowed';
+      }
+      return null;
    }
 }