+++ src/lib/st-select/st-select.module.ts
@@ -16,11 +16,11 @@ import { StDropdownMenuModule } from '../st-dropdown-menu/st-dropdown-menu.modul
 import { StSelectComponent } from './st-select';
 import { StCheckValidationsDirective } from './st-check-validations';
 import { StLabelModule } from '../st-label/st-label.module';
-import { StClickOutside } from '../directives/st-click-outside/st-click-outside.directive';
+import { StClickOutsideModule } from '../directives/st-click-outside/st-click-outside.module';
 
 @NgModule({
-   imports: [CommonModule, StDropdownMenuModule, FormsModule, ReactiveFormsModule, StLabelModule],
-   declarations: [StSelectComponent, StCheckValidationsDirective, StClickOutside],
+   imports: [CommonModule, StDropdownMenuModule, FormsModule, ReactiveFormsModule, StLabelModule, StClickOutsideModule],
+   declarations: [StSelectComponent, StCheckValidationsDirective],
    exports: [StSelectComponent, StCheckValidationsDirective]
 })
 export class StSelectModule {